backend/requirements.txt
pdfplumber==0.11.4
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==10.4.0
python-magic==0.4.27
pydantic==2.9.2
pytest==8.3.4
pytest-cov==6.0.0